+++ lib/utils/helpers/discount_percent.dart
@@ -1,5 +1,5 @@
-int? discountPercent(int? price, int? discount) {
+double? discountPercent(double? price, double? discount) {
   if (price == null || discount == null) return null;
 
-  return ((discount / price) * 100).toInt();
+  return (discount / price) * 100;
 }
